Renaming namespace within traits to match.
