+++ src/Controllers/Html/AbstractOwnerController.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace ResourceController\Controllers\Html;
 
-use ResourceController\Traits\Http\AssignOwner;
+use ResourceController\Traits\Controllers\AssignOwner;
 
 abstract class AbstractOwnerController extends AbstractModelController
 {
