+++ src/Controllers/Html/UserController.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace ResourceController\Controllers\Html;
 
 use Illuminate\Support\Facades\Hash;
-use ResourceController\Traits\Http\PaginateIndex;
+use ResourceController\Traits\Controllers\PaginateIndex;
 
 class UserController extends AbstractModelController
 {
