+++ src/Controllers/JsonApi/AbstractOwnerController.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace ResourceController\Controllers\JsonApi;
 
-use ResourceController\Traits\Http\AssignOwner;
+use ResourceController\Traits\Controllers\AssignOwner;
 
 abstract class AbstractOwnerController extends AbstractModelController
 {
