+++ src/Traits/Controllers/AssignOwner.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace ResourceController\Traits\Http;
+namespace ResourceController\Traits\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
